# Trigger memtable flush based on number of hidden entries scanned