--- a/db/column_family.cc
+++ b/db/column_family.cc
@@ -448,11 +448,11 @@ ColumnFamilyOptions SanitizeCfOptions(const ImmutableDBOptions& db_options,
     result.preclude_last_level_data_seconds = 0;
   }
 
-  if (read_only && result.memtable_tombstone_scan_limit != 0) {
+  if (read_only && result.tombstone_scan_flush_trigger != 0) {
     ROCKS_LOG_WARN(db_options.info_log.get(),
-                   "option memtable_tombstone_scan_limit is sanitized to "
+                   "option tombstone_scan_flush_trigger is sanitized to "
                    "0(disabled) for read only DB.");
-    result.memtable_tombstone_scan_limit = 0;
+    result.tombstone_scan_flush_trigger = 0;
   }
 
   return result;