@cbi42 2025-04-07
> not sure how you/the rocksdb team feels about it. But when adding new options, can we also include updating the C api to expose these options so that we don't have to retroactively add it? If that's not something you want to commit to, that's fine as well. I can create a follow up pr to expose this option since I want this feature.

We usually rely on external contributors for C bindings. I can help review the follow up PR.